restructured converter package
