+++ calendar_converter_root/calendar_converter/urls.py
@@ -14,12 +14,13 @@ Including another URLconf
     1. Import the include() function: from django.urls import include, path
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
-from django.contrib import admin
-from django.urls import path, include
+
 from django.conf import settings
 from django.conf.urls.static import static
+from django.contrib import admin
+from django.urls import include, path
 
 urlpatterns = [
-    path('admin/', admin.site.urls),
-    path("", include("converter.urls"))
+    path("admin/", admin.site.urls),
+    path("", include("file_converter.urls")),
 ] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
