+++ calendar_converter_root/file_converter/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class ConverterConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "file_converter"
