+++ calendar_converter_root/packages/converter/__init__.py
@@ -1,0 +1,3 @@
+from .transform import convert_file
+
+__all__ = ["convert_file"]
