--- converter/__init__.py
@@ -1,3 +1,0 @@
-from converter.src.converter import convert_file
-
-__all__ = ["convert_file"]
